AC
Here's the idea / thought path.

I started with two arrays: A[l][r], B[l][r] -- the scores for each player, when they have the first move on the range of [l, r]. Now, say, I'm the player A, and I pull from the front of the range -- the xs[l]. Let t be the range_sum(l, r). Then, t = xs[l] + B[l + 1][r] + alpha, which leads to alpha = t - xs[l] - B[l + 1][r]. Thus, A[l][r] = xs[l] + alpha = t - B[l + 1][r]. Similarly, if I pull from the back of the range, A[l][r] = t - B[l][r - 1]. The optimal strategy is to take the largest of the two numbers.

Next, we make an observation, that the A will be identical to B, and we only need one DP matrix.