AC \o/
Here's what's going on in that ../program.txt.

There are 3 nested loops:

for b in 106500..=123500 step 17 {
  for d in 2..=b {
    for e in 2..=b {
      if e * d - b == 0 {
        f = 0;
      }
      …
    }
  }
}

Thus, flag `f` is set iﬀ `e` is a divisor of `b`. Then, if flag f is
set, the sought value h gets incremented. Therefore, h is the number of
compound (non-prime) numbers in the range [106500, 123500].